# Avoid passing large arrays on heap, also use less memory