@kevaundray 2024-02-26
Can you comment on the difference this makes in geth?