@kevaundray 2024-03-08
I've made some changes and left some comments -- can merge once you look over them -- linter itself seems to be failing which is orthogonal to this PR (I ran it locally)